Update styling.md
Maybe with different wording or comment style, but I'd like to suggest calling out that /deep/ and ::shadow are not recommended here.  I know there is a section about it further up this page, but people (I, for example) tend to skim and look for answers quickly and may miss that section.  I had read this section and that section and came away wondering whether this was an exception or whether it was still a bad plan to use /deep/ ::shadow, even within a custom-style.

I generally assume that any good examples are *good* ways to do something unless something (a comment) calls out that they are a *bad*, counter-example.